Staging: vt6655: fix up U32 conversion
Commit 902d2411298a6e0b8a473a0163aa4bf6d5c8a073 converted U32 to u32
which you would think would be just fine.  Unfortunatly, it seems that
VIA only builds their code on a 32bit processor (which makes sense if
you think about it), but this doesn't work on x86-64.  So fix up the few
places where this really wanted to be an unsigned long width.

Cc: Charles Clément <caratorn@gmail.com>
Cc: Forest Bond <forest@alittletooquiet.net>,
Cc: Andres More <more.andres@gmail.com>

+++ drivers/staging/vt6655/device.h
@@ -424,7 +424,7 @@ typedef struct __device_info {
 
     CHIP_TYPE                   chip_id;
 
-    u32                         PortOffset;
+    unsigned long               PortOffset;
     DWORD                       dwIsr;
     u32                         memaddr;
     u32                         ioaddr;
